feat: add linux/arm64 support
linux/arm64 previously failed to build (amd64-only syscall asm and an x86-only
MAP_32BIT). Add it without cgo:
- asm_linux_arm64.s: _Syscall/_Syscall6 SVC trampolines, mirroring the amd64
  file (instruction sequence from Go's own arm64 syscall asm)
- split the arch-specific pieces into posix_linux_amd64.go and
  posix_linux_arm64.go: the _SYS_* syscall numbers (every one differs on arm64)
  and the Stat_t layout (arm64 uses the asm-generic struct stat, with Mode
  before Nlink and different padding)
- MAP_32BIT moves to the amd64 file (it does not exist on arm64)
- CI build and race matrices gain ubuntu-24.04-arm so the suite, including the
  Fstat ABI check, runs natively on arm64

Stays cgo-free and zero-dependency; the v1 import path is unchanged.
