+++ asm_linux_arm64.s
@@ -1,0 +1,62 @@
+// Copyright 2015 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the THIRD-PARTY-NOTICES.txt file.
+
+#include "textflag.h"
+
+//
+// System calls for arm64, Linux
+//
+// Args in R0-R5, syscall number in R8, return in R0. A return value in
+// [-4095, -1] is -errno (the CMN $4095 / BCC check below).
+
+// func _Syscall(trap, a1, a2, a3 uintptr) (r1, r2 uintptr, err syscall.Errno)
+TEXT ·_Syscall(SB),NOSPLIT,$0-56
+	CALL	runtime·entersyscall(SB)
+	MOVD	a1+8(FP), R0
+	MOVD	a2+16(FP), R1
+	MOVD	a3+24(FP), R2
+	MOVD	trap+0(FP), R8	// syscall entry
+	SVC
+	CMN	$4095, R0
+	BCC	ok
+	MOVD	$-1, R1
+	MOVD	R1, r1+32(FP)	// r1 = -1
+	MOVD	ZR, r2+40(FP)	// r2 = 0
+	NEG	R0, R0
+	MOVD	R0, err+48(FP)	// errno
+	CALL	runtime·exitsyscall(SB)
+	RET
+ok:
+	MOVD	R0, r1+32(FP)	// r1
+	MOVD	R1, r2+40(FP)	// r2
+	MOVD	ZR, err+48(FP)	// errno = 0
+	CALL	runtime·exitsyscall(SB)
+	RET
+
+// func _Syscall6(trap, a1, a2, a3, a4, a5, a6 uintptr) (r1, r2 uintptr, err syscall.Errno)
+TEXT ·_Syscall6(SB),NOSPLIT,$0-80
+	CALL	runtime·entersyscall(SB)
+	MOVD	a1+8(FP), R0
+	MOVD	a2+16(FP), R1
+	MOVD	a3+24(FP), R2
+	MOVD	a4+32(FP), R3
+	MOVD	a5+40(FP), R4
+	MOVD	a6+48(FP), R5
+	MOVD	trap+0(FP), R8	// syscall entry
+	SVC
+	CMN	$4095, R0
+	BCC	ok6
+	MOVD	$-1, R1
+	MOVD	R1, r1+56(FP)	// r1 = -1
+	MOVD	ZR, r2+64(FP)	// r2 = 0
+	NEG	R0, R0
+	MOVD	R0, err+72(FP)	// errno
+	CALL	runtime·exitsyscall(SB)
+	RET
+ok6:
+	MOVD	R0, r1+56(FP)	// r1
+	MOVD	R1, r2+64(FP)	// r2
+	MOVD	ZR, err+72(FP)	// errno = 0
+	CALL	runtime·exitsyscall(SB)
+	RET
