+++ const_linux.go
@@ -12,7 +12,6 @@ const (
 	MADV_DONTFORK    = syscall.MADV_DONTFORK
 	MADV_REMOVE      = syscall.MADV_REMOVE
 	MADV_UNMERGEABLE = syscall.MADV_UNMERGEABLE
-	MAP_32BIT        = syscall.MAP_32BIT
 	MAP_DENYWRITE    = syscall.MAP_DENYWRITE
 	MAP_EXECUTABLE   = syscall.MAP_EXECUTABLE
 	MAP_GROWSDOWN    = syscall.MAP_GROWSDOWN
