+++ shm_open_linux.go
@@ -68,8 +68,6 @@ func unlinkat(dirfd int, path string, flags int) (err error) {
 
 //goland:noinspection GoSnakeCaseUsage
 const (
-	prefix        = "/dev/shm/"
-	_SYS_OPENAT   = 257
-	_SYS_UNLINKAT = 263
-	_AT_FDCWD     = -0x64
+	prefix    = "/dev/shm/"
+	_AT_FDCWD = -0x64
 )
